pyproject.toml
[project]
name = "librarysite"
version = "0.1.0"
description = "Library site"
authors = [
    {name = "Selector0073"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (==5.2.6)",
    "beautifulsoup4 (==4.13.5)",
    "djangorestframework (==3.16.1)",
    "djangorestframework-simplejwt (==5.5.1)",
    "mimesis (==18.0.0)",
    "openpyxl (==3.1.5)",
    "progress (==1.6.1)",
    "python-decouple (==3.8)",
    "requests (==2.32.5)",
    "lxml (==6.0.2)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
